Coordinate in float. Problem with lookup.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/StructuredP2P@12822 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extensions/org/objectweb/proactive/extensions/structuredp2p/core/can/Coordinate.java
@@ -51,8 +51,8 @@ public class Coordinate implements Serializable {
      */
     public static Coordinate getMiddle(Coordinate min, Coordinate max) {
         // FIXME check with strings
-        int minL = Integer.parseInt(min.getValue());
-        int maxL = Integer.parseInt(max.getValue());
+        float minL = Float.parseFloat(min.getValue());
+        float maxL = Float.parseFloat(max.getValue());
 
         return new Coordinate("" + ((maxL + minL) / 2));
     }
@@ -168,9 +168,9 @@ public class Coordinate implements Serializable {
      *         coordinate argument; and a value greater than <code>0</code> if this coordinate is
      *         lexicographically greater than the coordinate argument.
      */
-    public int compareTo(Coordinate coord) {
-        int val = Integer.parseInt(coord.getValue());
-        int cur = Integer.parseInt(this.getValue());
+    public float compareTo(Coordinate coord) {
+        float val = Float.parseFloat(coord.getValue());
+        float cur = Float.parseFloat(this.getValue());
 
         return cur - val;
     }
